* commit 'e9210f0c0fc6f917331c1fd93724af02e2dfb55a':
  revert long variable conversion

+++ core/src/main/java/site/ycsb/Client.java
@@ -211,7 +211,7 @@ public final class Client {
    *
    * @throws IOException Either failed to write to output stream or failed to close it.
    */
-  private static void exportMeasurements(Properties props, long opcount, long runtime)
+  private static void exportMeasurements(Properties props, int opcount, long runtime)
       throws IOException {
     MeasurementsExporter exporter = null;
     try {
@@ -327,7 +327,7 @@ public final class Client {
     Thread terminator = null;
     long st;
     long en;
-    long opsDone;
+    int opsDone;
 
     try (final TraceScope span = tracer.newScope(CLIENT_WORKLOAD_SPAN)) {
 
@@ -408,18 +408,18 @@ public final class Client {
 
     final List<ClientThread> clients = new ArrayList<>(threadcount);
     try (final TraceScope span = tracer.newScope(CLIENT_INIT_SPAN)) {
-      long opcount;
+      int opcount;
       if (dotransactions) {
-        opcount = Long.parseLong(props.getProperty(OPERATION_COUNT_PROPERTY, "0"));
+        opcount = Integer.parseInt(props.getProperty(OPERATION_COUNT_PROPERTY, "0"));
       } else {
         if (props.containsKey(INSERT_COUNT_PROPERTY)) {
-          opcount = Long.parseLong(props.getProperty(INSERT_COUNT_PROPERTY, "0"));
+          opcount = Integer.parseInt(props.getProperty(INSERT_COUNT_PROPERTY, "0"));
         } else {
-          opcount = Long.parseLong(props.getProperty(RECORD_COUNT_PROPERTY, DEFAULT_RECORD_COUNT));
+          opcount = Integer.parseInt(props.getProperty(RECORD_COUNT_PROPERTY, DEFAULT_RECORD_COUNT));
         }
       }
       if (threadcount > opcount && opcount > 0){
-        threadcount = (int)opcount;
+        threadcount = opcount;
         System.out.println("Warning: the threadcount is bigger than recordcount, the threadcount will be recordcount!");
       }
       for (int threadid = 0; threadid < threadcount; threadid++) {
@@ -432,7 +432,7 @@ public final class Client {
           break;
         }
 
-        long threadopcount = opcount / threadcount;
+        int threadopcount = opcount / threadcount;
 
         // ensure correct number of operations, in case opcount is not a multiple of threadcount
         if (threadid < opcount % threadcount) {
